+++ src/translations/de/activeusers.php
@@ -20,5 +20,9 @@ return [
     'Message if nobody is logged in' => 'Nachricht, wenn niemand eingeloggt ist',
     'Inactive Time' => 'Inaktive Zeit',
     'Minutes after the user is assumed inactive' => 'Minuten ab wann der User als inaktiv angezeigt werden soll',
-    'minutes' => 'Minuten'
+    'minutes' => 'Minuten',
+    'User limit' => 'Anzahl Benutzer',
+    'Max. shown user' => 'Max. Anzahl der angezeigten Benutzer',
+    'Max. content height in px' => 'Max. Höhe des Widget-Inhalts',
+    'A scrollbar is added if the content height exceeds the height set (0 = unlimited)' => 'Ein Scrollbalken wird hinzugefügt, wenn der Inhalt die festgelegte Höhe übersteigt. (0 = unbegrenzt)'
 ];