+++ src/translations/en/activeusers.php
@@ -19,5 +19,10 @@ return [
     'Message' => 'Message',
     'Message if nobody is logged in' => 'Message if nobody is logged in',
     'Inactive Time' => 'Inactive Time',
-    'Minutes after the user is assumed inactive' => 'Minutes after the user is assumed inactive'
+    'Minutes after the user is assumed inactive' => 'Minutes after the user is assumed inactive',
+    'minutes' => 'minutes',
+    'User limit' => 'User limit',
+    'Max. shown user' => 'Max. shown user',
+    'Max. content height in px' => 'Max. content height in px',
+    'A scrollbar is added if the content height exceeds the height set (0 = unlimited)' => 'A scrollbar is added if the content height exceeds the height set (0 = unlimited)'
 ];